+++ Development/MyApp.swift
@@ -10,13 +10,7 @@ struct MyApp: App {
     var body: some Scene {
         WindowGroup {
             ZStack(alignment: .bottomTrailing) {
-#if os(watchOS) || os(tvOS)
-                ScrollView {
-                    ColorTestView()
-                }
-#else
                 ColorTestView()
-#endif
                 Text("Color v\(Color.version) © 2024 Kudit LLC").font(.caption).padding().backport.foregroundStyle(.white)
             }
         }
@@ -24,6 +18,9 @@ struct MyApp: App {
 }
 #if swift(>=5.9)
 // README examples
+#Preview("ColorTest") {
+    ColorTestView()
+}
 #Preview("Demo") {
     SimpleDemoView()
 }